Unify TextLink and NextLink
