+++ components/Dribbble/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Stack } from '../design-system/box';
 import { Heading, Text } from '../design-system/typography';
-import { TextLink } from '../design-system/text-link';
+import { TextLink } from '../TextLink';
 
 import { DribbbleGrid } from './DribbbleGrid';
 import { useFetchDribbbleShots } from './fetchDribbbleShots';
