+++ components/Mdx.tsx
@@ -11,11 +11,11 @@ import { Tag, Tags } from './design-system/tag';
 import { Box, Flex, Stack } from './design-system/box';
 import { Prose } from './design-system/prose';
 import { Card } from './design-system/card';
-import { TextLink } from './design-system/text-link';
+import { TextLink } from './TextLink';
 import { Text, Heading } from './design-system/typography';
 import { useTheme, spacing } from './design-system/theme';
 import { Divider } from './Divider';
-import { NextTextLink } from './NextTextLink';
+import { TextLink } from './TextLink';
 import { CodeEditor, StaticCode } from './CodeEditor';
 import { Table, TableHead, TableRow, TableCell } from './Table';
 
@@ -35,7 +35,7 @@ const DsComponents = {
   Tags,
   Text,
   TextLink,
-  NextTextLink,
+  TextLink,
   spacing
 };
