--- components/PostTemplate.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react';
-
-export const BlogTemplate = ({ children }: { children: ReactNode }) => {
-  return <div>{children}</div>;
-};
