+++ components/TextLink/index.ts
@@ -1,0 +1,2 @@
+export * from './TextLink';
+export * from './useLinkStyles';
