--- components/design-system/text-link/index.ts
@@ -1,1 +1,0 @@
-export * from './TextLink';
